backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==26.1.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
deprecation==2.1.0
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.4.0
django-csp==4.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.0.3
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
playwright==1.45.0
postgrest==2.27.2
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==11.1.0
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pypdfium2==5.7.0
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
realtime==2.27.2
redis==5.0.7
requests==2.33.1
rich==14.3.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
